+++ lib/cancellable.dart
@@ -4,6 +4,7 @@ export 'src/core/cancellable.dart';
 export 'src/core/cancellable_zone.dart';
 export 'src/exception/cancelled_exception.dart';
 export 'src/tools/cancellable_ext.dart';
+export 'src/tools/cancellable_future.dart';
 export 'src/tools/completer.dart';
 export 'src/tools/future_ext.dart';
 export 'src/tools/group.dart';